+++ CanPany-BE/CanPany.Application/Interfaces/Services/IPaymentService.cs
@@ -3,7 +3,7 @@ namespace CanPany.Application.Interfaces.Services;
 
 public interface IPaymentService
 {
-    Task<string> CreateTopUpAsync(string userId, decimal amount, string ip, CancellationToken ct = default);
+    Task<string> CreateTopUpAsync(string userId, long amount, string ip, CancellationToken ct = default);
     Task<(bool ok, string redirect)> HandleVnPayReturnAsync(IDictionary<string, string> queryParams, CancellationToken ct = default);
     Task<IEnumerable<Domain.Entities.Payment>> GetAllAsync();
 }
